refactor: Product Entity 타입 변경
- price: int -> Long 변경
- quantity: long -> Integer 변경

+++ src/main/java/org/phoenix/planet/dto/product/raw/Product.java
@@ -24,9 +24,9 @@ public class Product {
 
     private String name;
 
-    private int price;
+    private Long price;
 
-    private long quantity;
+    private Integer quantity;
 
     private String ecoDealStatus;
